Prepare composer integration
